pipeline/requirements.txt
# Core dependencies
feedparser>=6.0.10        # RSS/Atom feed parsing
requests>=2.31.0          # HTTP requests
python-dateutil>=2.8.2    # Date parsing
pytz>=2023.3              # Timezone handling
beautifulsoup4>=4.12.0    # HTML parsing
lxml>=4.9.3               # XML/HTML parsing (faster)

# Phase 2: AI scoring
openai>=1.12.0             # OpenAI API for embeddings and scoring
python-dotenv>=1.0.0       # Environment variable management
numpy>=1.24.0              # Vector operations for embeddings
scikit-learn>=1.3.0        # Cosine similarity calculations

# Content processing
readability-lxml>=0.8.1   # Content extraction

# Data handling
pyyaml>=6.0.1             # YAML config files

# Optional (for future phases)
# anthropic>=0.7.0           # Phase 2: Alternative AI
# schedule>=1.2.0            # Phase 3: Scheduling
# tweepy>=4.14.0             # Phase 3: Twitter integration
# openai>=1.0.0           # OpenAI API (Phase 2)
# python-dotenv>=1.0.0    # Environment variables
